// frontend-angular/src/proto/common.ts
export interface AreaEdge {
  toAreaId: string;
  travelMinutes: number;
}

export interface Area {
  id: string;
  name: string;
  isStation: boolean;
  neighbours: AreaEdge[];
}

export interface UserId {
  id: string;
}

function createBaseAreaEdge(): AreaEdge {
  return { toAreaId: "", travelMinutes: 0 };
}

export const AreaEdge: MessageFns<AreaEdge> = {
  encode(message: AreaEdge, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.toAreaId !== "") {
      writer.uint32(10).string(message.toAreaId);
    }
    if (message.travelMinutes !== 0) {
      writer.uint32(16).int32(message.travelMinutes);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AreaEdge {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAreaEdge();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.toAreaId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.travelMinutes = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AreaEdge {
    return {
      toAreaId: isSet(object.toAreaId) ? globalThis.String(object.toAreaId) : "",
      travelMinutes: isSet(object.travelMinutes) ? globalThis.Number(object.travelMinutes) : 0,
    };
  },

  toJSON(message: AreaEdge): unknown {
    const obj: any = {};
    if (message.toAreaId !== "") {
      obj.toAreaId = message.toAreaId;
    }
    if (message.travelMinutes !== 0) {
      obj.travelMinutes = Math.round(message.travelMinutes);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AreaEdge>, I>>(base?: I): AreaEdge {
    return AreaEdge.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AreaEdge>, I>>(object: I): AreaEdge {
    const message = createBaseAreaEdge();
    message.toAreaId = object.toAreaId ?? "";
    message.travelMinutes = object.travelMinutes ?? 0;
    return message;
  },
};

function createBaseArea(): Area {
  return { id: "", name: "", isStation: false, neighbours: [] };
}

export const Area: MessageFns<Area> = {
  encode(message: Area, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    if (message.name !== "") {
      writer.uint32(18).string(message.name);
    }
    if (message.isStation !== false) {
      writer.uint32(24).bool(message.isStation);
    }
    for (const v of message.neighbours) {
      AreaEdge.encode(v!, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Area {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseArea();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.isStation = reader.bool();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.neighbours.push(AreaEdge.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Area {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : "",
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      isStation: isSet(object.isStation) ? globalThis.Boolean(object.isStation) : false,
      neighbours: globalThis.Array.isArray(object?.neighbours)
        ? object.neighbours.map((e: any) => AreaEdge.fromJSON(e))
        : [],
    };
  },

  toJSON(message: Area): unknown {
    const obj: any = {};
    if (message.id !== "") {
      obj.id = message.id;
    }
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.isStation !== false) {
      obj.isStation = message.isStation;
    }
    if (message.neighbours?.length) {
      obj.neighbours = message.neighbours.map((e) => AreaEdge.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Area>, I>>(base?: I): Area {
    return Area.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Area>, I>>(object: I): Area {
    const message = createBaseArea();
    message.id = object.id ?? "";
    message.name = object.name ?? "";
    message.isStation = object.isStation ?? false;
    message.neighbours = object.neighbours?.map((e) => AreaEdge.fromPartial(e)) || [];
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}