// frontend-angular/src/proto/driver.ts
export interface RouteStop {
  sequence: number;
  areaId: string;
  isStation: boolean;
  arrivalOffsetMinutes: number;
}
// ...
export const RouteStop: MessageFns<RouteStop> = {
  encode(message: RouteStop, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.sequence !== 0) {
      writer.uint32(8).int32(message.sequence);
    }
    if (message.areaId !== "") {
      writer.uint32(18).string(message.areaId);
    }
    if (message.isStation !== false) {
      writer.uint32(24).bool(message.isStation);
    }
    if (message.arrivalOffsetMinutes !== 0) {
      writer.uint32(32).int32(message.arrivalOffsetMinutes);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RouteStop {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRouteStop();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.sequence = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.areaId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.isStation = reader.bool();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.arrivalOffsetMinutes = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RouteStop {
    return {
      sequence: isSet(object.sequence) ? globalThis.Number(object.sequence) : 0,
      areaId: isSet(object.areaId) ? globalThis.String(object.areaId) : "",
      isStation: isSet(object.isStation) ? globalThis.Boolean(object.isStation) : false,
      arrivalOffsetMinutes: isSet(object.arrivalOffsetMinutes) ? globalThis.Number(object.arrivalOffsetMinutes) : 0,
    };
  },

  toJSON(message: RouteStop): unknown {
    const obj: any = {};
    if (message.sequence !== 0) {
      obj.sequence = Math.round(message.sequence);
    }
    if (message.areaId !== "") {
      obj.areaId = message.areaId;
    }
    if (message.isStation !== false) {
      obj.isStation = message.isStation;
    }
    if (message.arrivalOffsetMinutes !== 0) {
      obj.arrivalOffsetMinutes = Math.round(message.arrivalOffsetMinutes);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<RouteStop>, I>>(base?: I): RouteStop {
    return RouteStop.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RouteStop>, I>>(object: I): RouteStop {
    const message = createBaseRouteStop();
    message.sequence = object.sequence ?? 0;
    message.areaId = object.areaId ?? "";
    message.isStation = object.isStation ?? false;
    message.arrivalOffsetMinutes = object.arrivalOffsetMinutes ?? 0;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}